Factor out code and add more arg parsing
