+++ gitlab-work-tracking/app/Cargo.toml
@@ -13,7 +13,7 @@ version = "0.1.0"
 
 [dependencies]
 anyhow = "1.0"
-clap = {version = "4.0", features = ["derive"]}
+clap = {version = "4.0", features = ["derive", "env", "cargo"]}
 dotenvy = "0.15"
 env_logger = "0.10"
 gitlab-work = {path = "../gitlab_work"}
